rename the `google_cloud.secret_manager` module to utils
